refactor: `01_stateless/example` 디렉토리 구조 변경 및 의존성 경로 수정
- `repository`, `data_source`, `model` 디렉토리를 `data` 디렉토리 하위로 이동했습니다.
- 이에 따라 관련 파일들의 import 경로를 수정했습니다.
- 일부 파일에서 불필요한 빈 줄을 제거했습니다.
