+++ lib/01_stateless/example/data/model/image_mode.dart
